requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
authlib==1.3.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
